# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/users/sync": {
      "post": {
        "description": "Call this right after Firebase login. Creates the user if first time, returns existing user otherwise. Check `isNewUser` to decide whether to show onboarding.",
        "operationId": "AuthController_syncUser",
        "parameters": [],
        "responses": {
          "201": {
            "description": "User synced successfully",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/SyncUserResponseDto"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid or missing Firebase token"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Sync Firebase user with database",
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/api/devices/pair": {
      "post": {
        "description": "Links a device by serial number to the current user. If the device was previously paired by this user, it re-pairs it.",
        "operationId": "DevicesController_pairDevice",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PairDeviceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Device paired successfully",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/DeviceDto"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Device already paired to another user"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Pair a new device",
        "tags": [
          "Devices"
        ]
      }
    },
    "/api/devices/my-device": {
      "get": {
        "description": "Returns the most recently paired device. Returns 404 if no device is paired.",
        "operationId": "DevicesController_getMyDevice",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Device retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/DeviceDto"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No device paired"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current user's paired device",
        "tags": [
          "Devices"
        ]
      }
    },
    "/api/devices/{id}/settings": {
      "patch": {
        "description": "Partially update device configuration. Send only the fields you want to change.",
        "operationId": "DevicesController_updateSettings",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDeviceSettingsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Settings updated",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/DeviceDto"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Device not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update device settings",
        "tags": [
          "Devices"
        ]
      }
    },
    "/api/devices/{id}/sync": {
      "post": {
        "description": "Update battery state, battery percentage and firmware version from device. Sets syncStatus to SYNCED.",
        "operationId": "DevicesController_syncDevice",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncDeviceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Device synced",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/DeviceDto"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Device not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Sync device telemetry",
        "tags": [
          "Devices"
        ]
      }
    },
    "/api/devices/{id}/status": {
      "get": {
        "description": "Lightweight status check — battery level, charge state, sync state, firmware.",
        "operationId": "DevicesController_getStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Status retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/DeviceStatusDto"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Device not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get device battery and sync status",
        "tags": [
          "Devices"
        ]
      }
    },
    "/api/devices/{id}/unpair": {
      "delete": {
        "description": "Removes the device record. The device can be re-paired later with the serial number.",
        "operationId": "DevicesController_unpairDevice",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Device unpaired",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/UnpairResponseDto"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Device not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Unpair device",
        "tags": [
          "Devices"
        ]
      }
    },
...
    "/api/coach/messages": {
      "post": {
        "description": "**Conversation flow:**\n\n1. First message: do NOT send `conversationId` → a new conversation is created.\n2. Follow-up messages: send the `conversationId` returned from the previous response to continue the same conversation.\n\n_Note: Currently returns mock responses. OpenAI integration coming soon._",
        "operationId": "CoachController_sendMessage",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMessageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Message sent, assistant response returned",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/SendMessageResponseDto"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Send a message to the AI coach",
        "tags": [
          "Coach"
        ]
      }
    },
    "/api/coach/conversations": {
      "get": {
        "description": "Returns conversations with the last message included for preview.",
        "operationId": "CoachController_getConversations",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Conversations retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/CoachConversationDto"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all coach conversations",
        "tags": [
          "Coach"
        ]
      }
    },
    "/api/coach/conversations/{id}/messages": {
      "get": {
        "description": "Returns messages in chronological order (oldest first).",
        "operationId": "CoachController_getMessages",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Messages retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/CoachMessageDto"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Conversation not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get full message history for a conversation",
        "tags": [
          "Coach"
        ]
      }
    }
  },
  "info": {
    "title": "Straight API",
    "description": "Backend API for the Straight posture correction mobile app. All endpoints except /api/users/sync require a Firebase Bearer token.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "UserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "uuid-v4"
          },
          "firebaseUid": {
            "type": "string",
            "example": "firebase-uid-abc123"
          },
          "email": {
            "type": "object",
            "example": "user@example.com"
          },
          "phoneNumber": {
            "type": "object",
            "example": "+905551234567"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "firebaseUid",
          "createdAt",
          "updatedAt"
        ]
      },
      "UserProfileDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "uuid-v4"
          },
          "userId": {
            "type": "string",
            "example": "uuid-v4"
          },
          "firstName": {
            "type": "object",
            "example": "Ayşe"
          },
          "lastName": {
            "type": "object",
            "example": "Yılmaz"
          },
          "gender": {
            "type": "string",
            "enum": [
              "MALE",
              "FEMALE",
              "NOT_SPECIFIED"
            ],
            "example": "FEMALE"
          },
          "dateOfBirth": {
            "type": "object",
            "example": "1995-06-15T00:00:00.000Z"
          },
          "heightCm": {
            "type": "object",
            "example": 168
          },
          "weightKg": {
            "type": "object",
            "example": 65.5
          },
          "onboardingCompleted": {
            "type": "boolean",
            "example": false
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "userId",
          "gender",
          "onboardingCompleted",
          "createdAt",
          "updatedAt"
        ]
      },
      "SyncUserResponseDto": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserDto"
          },
          "profile": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/UserProfileDto"
              }
            ]
          },
          "isNewUser": {
            "type": "boolean",
            "example": true,
            "description": "true if this is the first time the user logged in"
          }
        },
        "required": [
          "user",
          "profile",
          "isNewUser"
        ]
      },
...
      "DeviceDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "uuid-v4"
          },
          "userId": {
            "type": "string",
            "example": "uuid-v4"
          },
          "nickname": {
            "type": "object",
            "example": "My Straight Device"
          },
          "serialNumber": {
            "type": "string",
            "example": "STR-000123"
          },
          "firmwareVersion": {
            "type": "object",
            "example": "1.2.3"
          },
          "deviceMode": {
            "type": "string",
            "enum": [
              "CASUAL",
              "TRAINING"
            ],
            "example": "CASUAL"
          },
          "vibrationPower": {
            "type": "number",
            "example": 2,
            "description": "1–4 scale"
          },
          "vibrationSpeed": {
            "type": "string",
            "enum": [
              "FAST_SHORT",
              "MEDIUM_FAST",
              "MEDIUM_SLOW",
              "SLOW_LONG"
            ],
            "example": "MEDIUM_FAST"
          },
          "alertDelaySeconds": {
            "type": "number",
            "example": 5
          },
          "alertLevelDegrees": {
            "type": "number",
            "example": 30
          },
          "triggerLimitSeconds": {
            "type": "number",
            "example": 10
          },
          "batteryState": {
            "type": "string",
            "enum": [
              "EMPTY",
              "CRITICAL",
              "IDLE",
              "CHARGING",
              "FULL"
            ],
            "example": "IDLE"
          },
          "batteryPercentage": {
            "type": "object",
            "example": 85
          },
          "syncStatus": {
            "type": "string",
            "enum": [
              "UNSYNCED",
              "SYNCING",
              "SYNCED"
            ],
            "example": "UNSYNCED"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "userId",
          "serialNumber",
          "deviceMode",
          "vibrationPower",
          "vibrationSpeed",
          "alertDelaySeconds",
          "alertLevelDegrees",
          "triggerLimitSeconds",
          "batteryState",
          "syncStatus",
          "createdAt",
          "updatedAt"
        ]
      },
      "DeviceStatusDto": {
        "type": "object",
        "properties": {
          "batteryState": {
            "type": "string",
            "enum": [
              "EMPTY",
              "CRITICAL",
              "IDLE",
              "CHARGING",
              "FULL"
            ],
            "example": "CHARGING"
          },
          "batteryPercentage": {
            "type": "object",
            "example": 72
          },
          "syncStatus": {
            "type": "string",
            "enum": [
              "UNSYNCED",
              "SYNCING",
              "SYNCED"
            ],
            "example": "SYNCED"
          },
          "firmwareVersion": {
            "type": "object",
            "example": "1.3.0"
          }
        },
        "required": [
          "batteryState",
          "syncStatus"
        ]
      },
      "UnpairResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Device unpaired successfully"
          }
        },
        "required": [
          "message"
        ]
      },
      "PairDeviceDto": {
        "type": "object",
        "properties": {
          "serialNumber": {
            "type": "string",
            "example": "STR-000123",
            "description": "Device serial number"
          },
          "nickname": {
            "type": "string",
            "example": "My Straight Device"
          },
          "firmwareVersion": {
            "type": "string",
            "example": "1.2.3"
          }
        },
        "required": [
          "serialNumber"
        ]
      },
      "UpdateDeviceSettingsDto": {
        "type": "object",
        "properties": {
          "nickname": {
            "type": "string",
            "example": "My Device"
          },
          "deviceMode": {
            "type": "string",
            "enum": [
              "CASUAL",
              "TRAINING"
            ]
          },
          "vibrationPower": {
            "type": "number",
            "example": 2,
            "minimum": 1,
            "maximum": 4
          },
          "vibrationSpeed": {
            "type": "string",
            "enum": [
              "FAST_SHORT",
              "MEDIUM_FAST",
              "MEDIUM_SLOW",
              "SLOW_LONG"
            ]
          },
          "alertDelaySeconds": {
            "type": "number",
            "example": 5,
            "description": "Alert delay in seconds"
          },
          "alertLevelDegrees": {
            "type": "number",
            "example": 30,
            "description": "Alert level in degrees"
          },
          "triggerLimitSeconds": {
            "type": "number",
            "example": 10,
            "description": "Trigger limit in seconds"
          }
        }
      },
      "SyncDeviceDto": {
        "type": "object",
        "properties": {
          "batteryState": {
            "type": "string",
            "enum": [
              "EMPTY",
              "CRITICAL",
              "IDLE",
              "CHARGING",
              "FULL"
            ]
          },
          "batteryPercentage": {
            "type": "number",
            "example": 85,
            "minimum": 0,
            "maximum": 100
          },
          "firmwareVersion": {
            "type": "string",
            "example": "1.3.0"
          }
        }
      },
...
      "CoachMessageDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "uuid-v4"
          },
          "conversationId": {
            "type": "string",
            "example": "uuid-v4"
          },
          "role": {
            "type": "string",
            "enum": [
              "USER",
              "ASSISTANT"
            ],
            "example": "ASSISTANT"
          },
          "content": {
            "type": "string",
            "example": "Great question! To improve your posture while sitting, keep your feet flat on the floor."
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "conversationId",
          "role",
          "content",
          "createdAt"
        ]
      },
      "SendMessageResponseDto": {
        "type": "object",
        "properties": {
          "conversationId": {
            "type": "string",
            "example": "uuid-v4",
            "description": "Use this ID for follow-up messages"
          },
          "message": {
            "$ref": "#/components/schemas/CoachMessageDto"
          }
        },
        "required": [
          "conversationId",
          "message"
        ]
      },
      "LastMessageDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "uuid-v4"
          },
          "role": {
            "type": "string",
            "enum": [
              "USER",
              "ASSISTANT"
            ],
            "example": "USER"
          },
          "content": {
            "type": "string",
            "example": "How can I improve my posture?"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "role",
          "content",
          "createdAt"
        ]
      },
      "CoachConversationDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "uuid-v4"
          },
          "userId": {
            "type": "string",
            "example": "uuid-v4"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "messages": {
            "description": "Last message in conversation",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LastMessageDto"
            }
          }
        },
        "required": [
          "id",
          "userId",
          "createdAt",
          "messages"
        ]
      },
      "SendMessageDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "example": "How can I improve my posture while sitting?"
          },
          "conversationId": {
            "type": "string",
            "example": "uuid-of-existing-conversation",
            "description": "If provided, continues existing conversation"
          }
        },
        "required": [
          "content"
        ]
      }
    }
  }
}